feat ObjectOctreeNode

+++ Source/GameEngine/Graphics/Mesh/Include/Mesh.h
@@ -32,7 +32,7 @@ namespace SpaceGameEngine
 		void Release();
 		~MeshComponent();
 
-		void InitFromMemory(int VertexSize, int IndexSize, DefaultVertex* pVertices, WORD* pIndices);	//创建顶点缓存和索引缓存,VertexSize为顶点数,IndexSize为索引数（索引的多少）
+		void InitFromMemory(int VertexSize, int IndexSize, DefaultVertex* pVertices, unsigned int* pIndices);	//创建顶点缓存和索引缓存,VertexSize为顶点数,IndexSize为索引数（索引的多少）
 
 		void InitFromFile(const std::string& filename, int mode = 1);			//从文件读取Mesh
 		void Start();
